requirements.txt
streamlit
pandas
openai
matplotlib
gspread
google-auth
